Docker : Pin yum-installed package versions
We make use of yum-versionlock to pin the versions of the packages we will
install via yum. This means that repeating a build of the docker image
should produce a stable container even if newer versions of installed
packages are available.

We copy in a `versionlock.list` from the repository before installing
any packages. In order to reset the locks, you can simply remove any or all
of the entires in `yum-versionlock.list` before building the image.

To avoid storing pins for *all* packages, we take a snapshot of the
package list from the base image at the beginning of the build. Install
our packages, then only add locks for packages installed since.

We update the lock file in the container at the end of the build after
each run so that it can be copied out of the container should there be
any new locks to commit to source control.