Build : Allow the Docker image to be built with build.py
The 'build' project can now be specified, which will build the Docker
image from the local checkout, and optionally push this to Dockerhub
(if `--upload 1` is set).

Builds of the other project now use the published images (or their local
equivalents) instead of directly building the image each time.

Important changes :

 - `--docker` is now `--use-docker` to avoid ambiguity when building the
   docker image.

 - `--build-env` overrides the dockerhub image tag to use for builds of
   the 'gaffer' and 'dependencies' projects.

 - `--tag-as` defines the tag used to tag the image in image builds.

 - `--no-cache` Can be set to disable docker's layer cache during image
   builds (see README.md)

 - `--update-version-locks` and --new-only` Allow the yum package versions
   to be updated during an image build.