Docker : Remove build.py from built images
Storing build.py in images creates a backwards compatability problem as
newer build.py's would need to maintain compatability with old build.py's
invocation to allow building in older environments.

Now the image is simply a 'place in which the current build.py can run',
which should make swapping out the image to a plethora of build environments
much easier.