Dockerfile : Make openssl11 discoverable by Python
Python's `configure` script looks for OpenSSL in a variety of fallback paths if it can't find it via `pkg-config`. `openssl11-devel` doesn't install to any of these paths, and doesn't install libs and headers under a common root suitable for configure's `--with-openssl` flag, so we symlink into the first of the fallback paths.

We remove `openssl-devel` to prevent Python from finding it via `pkg-config`. This is of no use as Python 3.10 dropped support for OpenSSL 1.0 and prevents the above discovery of `openssl11-devel`.