build.py : Use `podman`
This is what we're using to build the container in CI, and is generally a bit nicer than Docker. One big benefit is that you don't need to run a daemon to use `podman`, particularly since the Docker daemon breaks the licensing for a certain piece of software that might be a little bit important at the moment.